+++ src/main/java/de/tomalbrc/cameraobscura/render/renderer/ImageRenderer.java
@@ -1,12 +1,16 @@
 package de.tomalbrc.cameraobscura.render.renderer;
 
+import me.drex.nem.config.ModConfig;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.phys.Vec3;
 
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.List;
+import java.util.concurrent.ForkJoinPool;
 
 public class ImageRenderer extends AbstractRenderer<int[]> {
+
+    private static final ForkJoinPool pool = new ForkJoinPool(ModConfig.getInstance().renderThreadCount);
+
     public ImageRenderer(LivingEntity entity, int width, int height, int renderDistance) {
         super(entity, width, height, renderDistance);
     }
@@ -16,12 +20,17 @@ public class ImageRenderer extends AbstractRenderer<int[]> {
 
         int[] pixels = new int[width * height];
 
-        this.castRays(this.entity).parallelStream().forEach(ray -> {
-            int color = raytracer.trace(eyes, ray.direction());
-            int x = width - ray.xOrigin() - 1;
-            int y = ray.yOrigin();
-            pixels[x + width * y] = color;
-        });
+        List<Ray> rays = this.castRays(this.entity);
+        pool.submit(() -> {
+            // https://stackoverflow.com/questions/21163108/custom-thread-pool-in-java-8-parallel-stream
+            rays.parallelStream().forEach(ray -> {
+                int color = raytracer.trace(eyes, ray.direction());
+                int x = width - ray.xOrigin() - 1;
+                int y = ray.yOrigin();
+                pixels[x + width * y] = color;
+            });
+        }).join();
+
 
         return pixels;
     }