+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `renderThreadCount` config option
+
+### Fixed
+- Some potential concurrency issues
+
 ## [1.2.0] - 2025-08-18
 ### Added
 - Computer block rotation